完成大部分，差组队建议

+++ graduation/src/main/java/com/cslg/graduation/dao/ScoreMapper.java
@@ -21,7 +21,7 @@ public interface ScoreMapper {
     public double findDailyScoreByUsername(String username, Date time);
 
     // 根据学号和时间获取用户该日排名
-    public double findRankByUsername(String username, Date time);
+    public int findRankByUsername(String username, Date time);
 
     // 新增一条积分
     public void insertScore(Score score);
